Fix another wire bug
The problem here seems to be the fact that when a wire was destroyed the
dragging property was not reset, which meant that the next time a tip entered a
socket DropArea, the DropArea would not pick it up. Also separated the dragging
of the start and the end tips (previously they would both go active when one of
them did because they both depend on the shared `root.dragging` property).