Refactor the wires so that their lifetime is controlled from C++
This fixes an edge-case where QML's internal state would be corrupted when a
socket that was the QObject parent of a wire was deleted. Regrettably, this also
introduced an old bug that causes the DropArea's to be deactivated when their
wire was deleted (some old code was written to handle this in Wire.qml, but it
did not work in the new code so it was removed).