Ensure that all sockets enter the undo/redo system
Previously this was not the case because SocketModel::setTemplate() would run
when the nodes were created and create extra sockets upon a call to
DeleteNode::undo(). But now it is only run if the node has not been deleted
before, and when it does run it emits XCom::requestCreateSocket() instead of
directly adding it to the model.